vasi-sync: Move sync abstractions to crate-level module
We'll want to reuse these for additional modules in this crate.
